refactor(ui): rename ISessions.ts to ISession.ts
